+++ src/hooks/useGameBoard.ts
@@ -8,6 +8,7 @@ import {
   selectStageColors,
 } from "../utils/gameLogic";
 import { useAnimationSpeed } from "../contexts/AnimationSpeedContext"; // AnimationSpeedContext をインポート
+import { saveGameHistory } from "../utils/saveGameHistory";
 
 // Difficulty 型を定義
 type Difficulty = "easy" | "medium" | "hard";
@@ -153,23 +154,7 @@ const useGameBoard = (initialDifficulty: Difficulty) => {
   // ゲームオーバー時に最高クリアステージと過去10回の到達ステージ履歴を更新
   useEffect(() => {
     if (isGameOver) {
-      const lastClearedStage = Math.max(0, stage - 1); // 現在のステージの1つ前が最後にクリアしたステージ (最低0)
-
-      // 過去10回の到達ステージ履歴の更新
-      const storedHistory = localStorage.getItem("stageHistory");
-      let history: number[] = storedHistory ? JSON.parse(storedHistory) : [];
-
-      // 新しいステージを追加
-      history.push(lastClearedStage);
-
-      // 履歴が10件を超えたら古いものから削除
-      if (history.length > 10) {
-        history = history.slice(history.length - 10);
-      }
-
-      // 更新した履歴をlocalStorageに保存
-      localStorage.setItem("stageHistory", JSON.stringify(history));
-      console.log("Updated stage history:", history);
+      saveGameHistory(stage);
     }
   }, [isGameOver, stage]); // highestStageCleared も依存配列に残す
